feat: #5 Game -> FastAPI로 컨트롤러 및 서비스 코드 이전
