+++ src/main/java/com/Eungbyeonmuga/RookieKBO/domain/game/service/GameServiceImpl.java
@@ -31,149 +31,10 @@ import java.util.List;
 public class GameServiceImpl implements GameService {
 
     private final GameRepository gameRepository;
-    private final FastAPIClient fastAPIClient;
     private final GameMapper gameMapper;
     private final GameTeamService gameTeamService;
-    private final TeamService teamService;
-
-    // TODO: 코드 리팩토링 필요
-    // 우천 취소와 같은 기타 상황을 위한 기본 Game 셋팅
-
-    /***
-     * 사용 X KBO 리스트 기반인데, 데이터가 정확하지 않아서 캘린더 기반으로 변경
-     * @param year
-     * @param month
-     */
-    @Override
-    public GameResponse.GetGamesByYearAndMonth createGameBaseInfoForMonthFromKBOList(Integer year, Integer month) {
-
-        FastAPIRequest.GetGames getGamesRequest = FastAPIRequest.GetGames
-                .builder()
-                .year(year)
-                .month(month)
-                .build();
-
-        // FastAPI 서버에서 클로링 후 Game 생성 및 GameTeam 생성
-        FastAPIResponse.GetGames gameSummaries = fastAPIClient.getGames(getGamesRequest);
-        System.out.println(gameSummaries);
-        // 정규 시즌 생성
-        for (FastAPIResponse.GameSummary gameSummary : gameSummaries.getGameSummariesInRegularSeason()) {
-            System.out.println(gameSummary);
-            List<Integer> timeIntegerList = parseTimeStringToInteger(gameSummary.getDate(), gameSummary.getTime());
-            LocalDateTime startDateTime = convertToStartDateTime(year, month, timeIntegerList.get(0), timeIntegerList.get(1), timeIntegerList.get(2));
-            Status status;
-            if (gameSummary.getNote().contains("취소")) {
-                status = Status.CANCEL;
-            } else {
-                if (LocalDateTime.now().plusMinutes(5).isAfter(startDateTime)) {
-                    status = Status.OVER;
-                } else {
-                    status = Status.PREPARING;
-                }
-            }
-            Game newGame = gameMapper.toNewGame(Season.REGULAR_SEASON, startDateTime, gameSummary.getPlace(), gameSummary.getHomeScore(), gameSummary.getAwayScore(), status, gameSummary.getNote());
-            gameRepository.save(newGame);
-            Team homeTeam = teamService.findTeamByName(gameSummary.getHomeTeam());
-            Team awayTeam = teamService.findTeamByName(gameSummary.getAwayTeam());
-
-            gameTeamService.createGameTeam(newGame, homeTeam, awayTeam);
-        }
-        // 포스트 시즌 생성
-        for (FastAPIResponse.GameSummary gameSummary : gameSummaries.getGameSummariesInPostSeason()) {
-            System.out.println(gameSummary);
-            List<Integer> timeIntegerList = parseTimeStringToInteger(gameSummary.getDate(), gameSummary.getTime());
-            LocalDateTime startDateTime = convertToStartDateTime(year, month, timeIntegerList.get(0), timeIntegerList.get(1), timeIntegerList.get(2));
-            Status status;
-            if (gameSummary.getNote().contains("취소")) {
-                status = Status.CANCEL;
-            } else {
-                if (LocalDateTime.now().plusMinutes(5).isAfter(startDateTime)) {
-                    status = Status.OVER;
-                } else {
-                    status = Status.PREPARING;
-                }
-            }
-            Game newGame = gameMapper.toNewGame(Season.POST_SEASON, startDateTime, gameSummary.getPlace(), gameSummary.getHomeScore(), gameSummary.getAwayScore(), status, gameSummary.getNote());
-            gameRepository.save(newGame);
-            Team homeTeam = teamService.findTeamByName(gameSummary.getHomeTeam());
-            Team awayTeam = teamService.findTeamByName(gameSummary.getAwayTeam());
-
-            gameTeamService.createGameTeam(newGame, homeTeam, awayTeam);
-        }
-
-        return null;
-    }
-
-    // TODO: 코드 리팩토링 필요
-
-    /***
-     * KBO 홈페이지의 경기 일정 캘린더 기준 크롤링 데이터 패칭
-     * @param year
-     * @param month
-     */
-    // 우천 취소와 같은 기타 상황을 위한 기본 Game 셋팅
-    @Override
-    public GameResponse.GetGamesByYearAndMonth createGameBaseInfoForMonthFromKBOCalendar(Integer year, Integer month) {
-
-        List<Game> games = gameRepository.findByStartDate(convertToStartDateTime(year, month, 1, 0, 0));
-        FastAPIRequest.GetGameSummariesOnCalendar getGameSummariesOnCalendar = FastAPIRequest.GetGameSummariesOnCalendar
-                .builder()
-                .year(year)
-                .month(month)
-                .build();
 
 
-        // FastAPI 서버에서 클로링 후 Game 생성 및 GameTeam 생성
-        FastAPIResponse.GetGameSummariesOnCalendar gameSummariesOnCalendar = fastAPIClient.getGameSummariesOnCalendar(getGameSummariesOnCalendar);
-        System.out.println(gameSummariesOnCalendar);
-        // 정규 시즌 생성
-        for (FastAPIResponse.GameSummaryOnCalendar gameSummaryOnCalendar : gameSummariesOnCalendar.getGameSummariesOnCalendarInRegularSeason()) {
-            System.out.println(gameSummaryOnCalendar);
-
-            String stadium = teamService.findTeamByName(gameSummaryOnCalendar.getHomeTeam()).getStadium();
-
-            // LocalDate에 시간 정보를 추가하여 LocalDateTime으로 변환
-            LocalDateTime startDate = convertStringToLocalDateTime(gameSummaryOnCalendar.getDate()); // 자정 시간으로 변환
-
-            Status status = Status.fromKorean(gameSummaryOnCalendar.getGameStatus());
-
-            String note = "-";
-            if (status == Status.CANCEL) {
-                note = "경기 취소";
-            }
-
-            Game newGame = gameMapper.toNewGame(Season.REGULAR_SEASON, startDate, stadium, gameSummaryOnCalendar.getHomeScore(), gameSummaryOnCalendar.getAwayScore(), status, note);
-            gameRepository.save(newGame);
-            Team homeTeam = teamService.findTeamByName(gameSummaryOnCalendar.getHomeTeam());
-            Team awayTeam = teamService.findTeamByName(gameSummaryOnCalendar.getAwayTeam());
-
-            gameTeamService.createGameTeam(newGame, homeTeam, awayTeam);
-        }
-        // 포스트 시즌 생성
-        for (FastAPIResponse.GameSummaryOnCalendar gameSummaryOnCalendar : gameSummariesOnCalendar.getGameSummariesOnCalendarInPostSeason()) {
-            System.out.println(gameSummaryOnCalendar);
-
-            String stadium = teamService.findTeamByName(gameSummaryOnCalendar.getHomeTeam()).getStadium();
-
-            // LocalDate에 시간 정보를 추가하여 LocalDateTime으로 변환
-            LocalDateTime startDate = convertStringToLocalDateTime(gameSummaryOnCalendar.getDate()); // 자정 시간으로 변환
-
-            Status status = Status.fromKorean(gameSummaryOnCalendar.getGameStatus());
-
-            String note = "-";
-            if (status == Status.CANCEL) {
-                note = "경기 취소";
-            }
-
-            Game newGame = gameMapper.toNewGame(Season.POST_SEASON, startDate, stadium, gameSummaryOnCalendar.getHomeScore(), gameSummaryOnCalendar.getAwayScore(), status, note);
-            gameRepository.save(newGame);
-            Team homeTeam = teamService.findTeamByName(gameSummaryOnCalendar.getHomeTeam());
-            Team awayTeam = teamService.findTeamByName(gameSummaryOnCalendar.getAwayTeam());
-
-            gameTeamService.createGameTeam(newGame, homeTeam, awayTeam);
-        }
-        return null;
-    }
 
     @Override
     public GameResponse.GetGamesByDate getGamesByDate(String date) {
@@ -202,42 +63,6 @@ public class GameServiceImpl implements GameService {
                     return gameMapper.toGameInfo(game, homeTeam, awayTeam);
                 })
                 .toList());
-    }
-
-    @Override
-    @Transactional
-    public GameResponse.UpdateGamesDetailByDate updateGamesDetailByDate(Integer year, Integer month, Integer day) {
-
-        FastAPIRequest.GetGameDetailByDate request = FastAPIRequest.GetGameDetailByDate.builder()
-                .year(year)
-                .month(month)
-                .day(day)
-                .build();
-
-        FastAPIResponse.GetGameDetailByDate getGameDetailByDate = fastAPIClient.getGameDetailByDate(request);
-
-        for (FastAPIResponse.GameDetail gameDetail : getGameDetailByDate.getGameDetails()) {
-
-            Game game = gameRepository.findByStartDateTimeAndTeams(LocalDateTime.parse(gameDetail.getStartDateTime()), gameDetail.getHomeTeam(), gameDetail.getAwayTeam());
-            System.out.println(game);
-            if (game == null) {
-                return GameResponse.UpdateGamesDetailByDate.builder()
-                        .resultMessage("기본 경기 데이터가 없습니다. 월 단위 경기 기본 데이터를 생성한 후 다시 호출하세요.")
-                        .build();
-            }
-
-            game.updateHomeScores(gameDetail.getHomeScores());
-            game.updateAwayScores(gameDetail.getAwayScores());
-            game.updateHomeRHEB(gameDetail.getHomeRHEB());
-            game.updateAwayRHEB(gameDetail.getAwayRHEB());
-            game.updateStatus(Status.fromKorean(gameDetail.getGameStatus()));
-            game.updateStartDateTime(LocalDateTime.parse(gameDetail.getStartDateTime()));
-        }
-
-        System.out.println(getGameDetailByDate);
-        return GameResponse.UpdateGamesDetailByDate.builder()
-                .resultMessage("데이터 업데이트 성공")
-                .build();
     }
 
     @Override
